+++ src/main/scala/com/atomist/tree/ContainerTreeNodeImpl.scala
@@ -3,11 +3,11 @@ package com.atomist.tree
 /**
   * Simple implementation of ContainerTreeNode allowing fields to be added
   * @param nodeName node name
-  * @param nodeType node type
+  * @param nodeTags node type
   */
 class ContainerTreeNodeImpl(
                              val nodeName: String,
-                             override val nodeType: Set[String])
+                             override val nodeTags: Set[String])
   extends ContainerTreeNode {
 
   def this(nodeName: String, nt: String) =
@@ -21,11 +21,11 @@ class ContainerTreeNodeImpl(
 
   override def childNodeNames: Set[String] = kids.map(_.nodeName).toSet
 
-  override def childNodeTypes: Set[String] = kids.flatMap(_.nodeType).toSet
+  override def childNodeTypes: Set[String] = kids.flatMap(_.nodeTags).toSet
 
   override def childrenNamed(key: String): Seq[TreeNode] = kids.filter(_.nodeName.equals(key))
 
   override def value: String = s"$nodeName:${kids.mkString(",")}"
 
-  override def toString = s"$nodeName:$nodeType(${kids.size})"
+  override def toString = s"$nodeName:$nodeTags(${kids.size})"
 }