+++ src/main/scala/com/atomist/tree/pathexpression/ExpressionEngine.scala
@@ -22,7 +22,7 @@ object ExecutionResult {
   }
 
   def show(n: TreeNode): String = {
-    s"${n.nodeName}:${n.nodeType}"
+    s"${n.nodeName}:${n.nodeTags}"
   }
 
 }